read only the channels a postprocessor consumes, and stop widening labellings
Found while sizing the first full scoring job, which would have been killed part-way
through a ten-hour run.

`Artifact.read` returned every channel and the consumer sliced afterwards. For the
zebrafish doublecube -- 7.08 gigavoxels, six affinity channels -- that is 85 GB where
`cc_threshold` reads three and needs 42 GB. Postprocessors now declare `reads_channels()`
and the read honours it.

`cc_threshold` also widened the components pass's uint32 output to int64, which at that
size holds both copies at once: 85 GB where 28 GB will do. The metrics factorise ids
rather than assuming a width, so the widening bought nothing; verified that panoptic
quality and VOI are identical across uint16, int32, uint32 and int64. `identity` likewise
returns the array as read instead of casting it.

Together these take the projected peak for that volume from ~276 GB to ~205 GB, which is
the difference between fitting the reservation and not.

+++ src/postprocess/labellings.py
@@ -33,7 +33,9 @@ class Identity(BasePostprocess):
                 f"identity was handed a floating-point array (dtype {array.dtype}); a labelling "
                 "must be integral, or the ids are not ids. Check the artifact's `kind`."
             )
-        return array.astype(np.int64, copy=False)
+        # `copy=False` and no widening: a labelling read straight from an artifact is already the
+        # right thing, and at these volumes an unnecessary cast is tens of gigabytes.
+        return array
 
 
 @PostprocessRegistry.register("argmax")
